+++ CHANGES.md
@@ -6,6 +6,8 @@
 
 - Added function ijsonLoad() to load faster and save tags conns to json
 
+- Fixed bug in convergence and divergence connectivity -- now fixed number and no self conns
+
 - Fixed bug by converting tuples to strings when saving Matlab; now converts to lists
 
 - Fixed bug use int as argument to np.full in fixedInterval variable